+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 ## __WORK IN PROGRESS__
 * (@Apollon77) Allows only numbers for ts and tc fields in state when provided for setState
+* (@foxriver76) Added objects warn limit per instance
 
 ## 7.0.7 (2025-04-17) - Lucy
 * (@foxriver76) fixed the edge-case problem on Windows (if adapter calls `readDir` on single file)